✨ Create Footer component
