+++ client/src/AppRoutes.jsx
@@ -1,18 +1,10 @@
 import { Route, Routes } from 'react-router-dom';
-import Layout from './Layout.tsx';
+import Layout from './layout/Layout.tsx';
 
 const AppRoutes = () => {
   return (
     <Routes>
-      <Route
-        path="/"
-        element={
-          <Layout>
-            <main className="bg-gradient-to-b from-sky-500/[0.5] to-neutral-900 rounded-lg p-4 overflow-hidden"></main>
-            <div className="p-4">Player</div>
-          </Layout>
-        }
-      />
+      <Route path="/" element={<Layout>Main page</Layout>} />
     </Routes>
   );
 };
