+++ client/src/layout/Footer/Footer.module.scss
@@ -1,0 +1,3 @@
+.footer {
+  @apply px-8 pt-5 pb-10 border-t border-white/[0.10];
+}
