+++ client/src/layout/Footer/Footer.tsx
@@ -1,0 +1,7 @@
+import styles from './Footer.module.scss';
+
+const Footer = () => {
+  return <footer className={styles.footer}>Footer</footer>;
+};
+
+export default Footer;
