+++ client/src/layout/Layout.tsx
@@ -1,0 +1,26 @@
+import { ReactNode } from 'react';
+import styles from '../AppRoutes.module.scss';
+import Nav from './Nav/Nav';
+import Library from '../features/library/Library';
+import Footer from './Footer/Footer.tsx';
+
+interface LayoutProps {
+  children: ReactNode;
+}
+
+const Layout = ({ children }: LayoutProps) => {
+  return (
+    <div className={styles.app}>
+      <Nav />
+      <Library />
+      <main className="bg-gradient-to-b from-sky-500/[0.5] to-neutral-900 rounded-lg px-5 py-4 overflow-hidden">
+        {children}
+        <Footer />
+      </main>
+
+      <div className="p-4">Player</div>
+    </div>
+  );
+};
+
+export default Layout;
